switched to gpt-3.5-turbo added vectorization
I switched most of the api calls to be gpt-3.5-turbo to reduce costs, works the same, if not better for the summarizations and sentence making. Started to make a memory system that uses vectorization to find relevent past sentenced. The main.py script runs the basic search, test run 2 and 3 are attempts to produce coherent sentences, test run 4 saves the final thought(sentence) to the memory. Looping through this doesn't work well because of the global num.